build: Add bootloader & nandwrite to GRANDFATHERED_ALL_PREBUILT list
Starting from ICS, ALL_PREBUILT usage is deprecated.
Compilation exists if any of module/s uses ALL_PREBUILT.
As bootloader & nandwrite modules are compiled in
different way and to be in sync with ABS to get rid
of the compilation errors, adding these modules to
GRANDFATHERED_ALL_PREBUILT list.

Change-Id: I411831a5abc7c1a4e3e5783754c68c697e911b38
CRs-Fixed: 314291